assistant-core/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pcerson-assistant-core"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

